requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.1.0
zope.interface==5.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
drf-yasg==1.17.1
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
uritemplate==3.0.1
urllib3==1.25.9
